+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New card "Ballus, Dogfight Enforcer Q"
 - New card "Split-Head Hydroturtle Q"
+- New card "Bladerush Skyterror Q"
 
 ## [v2.3] - 09/04/2022